backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
openai>=1.30.1
praw==7.7.1
pytrends==4.9.2
python-dotenv==1.0.0 